[refactor] move Message into Messages sub-domain
