+++ app/Domain/Messages/Message.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace App\Domain;
+namespace App\Domain\Messages;
+
+use App\Domain\IEventPublisher;
 
 class Message {
     public static function publish(IEventPublisher $eventPublisher, $messageContent) {
